Revert "2.16.7-SNAPSHOT to fix bybit get position"
This reverts commit 7a57cf6d5c3e0c74f7fbb479876a3a9da880dac4.
