--- bybitlinear-api/src/main/java/io/contek/invoker/bybitlinear/api/rest/user/GetPosition.java
@@ -1,66 +1,0 @@
-package io.contek.invoker.bybitlinear.api.rest.user;
-
-import com.google.common.collect.ImmutableList;
-import io.contek.invoker.bybitlinear.api.common._Position;
-import io.contek.invoker.bybitlinear.api.rest.common.RestResponse;
-import io.contek.invoker.commons.actor.IActor;
-import io.contek.invoker.commons.actor.ratelimit.TypedPermitRequest;
-import io.contek.invoker.commons.rest.RestContext;
-import io.contek.invoker.commons.rest.RestMethod;
-import io.contek.invoker.commons.rest.RestParams;
-
-import javax.annotation.concurrent.NotThreadSafe;
-import java.util.List;
-import java.util.Objects;
-
-import static io.contek.invoker.bybitlinear.api.ApiFactory.RateLimits.ONE_REST_PRIVATE_POSITION_READ_REQUEST;
-import static io.contek.invoker.bybitlinear.api.rest.user.GetPosition.Response;
-import static io.contek.invoker.commons.rest.RestMethod.GET;
-
-@NotThreadSafe
-public final class GetPosition extends UserRestRequest<Response> {
-
-  private String symbol;
-
-  GetPosition(IActor actor, RestContext context) {
-    super(actor, context);
-  }
-
-  public GetPosition setSymbol(String symbol) {
-    this.symbol = symbol;
-    return this;
-  }
-
-  @Override
-  protected RestMethod getMethod() {
-    return GET;
-  }
-
-  @Override
-  protected String getEndpointPath() {
-    return "/private/linear/position/list";
-  }
-
-  @Override
-  protected RestParams getParams() {
-    RestParams.Builder builder = RestParams.newBuilder();
-
-    Objects.requireNonNull(symbol);
-    builder.add("symbol", symbol);
-
-    return builder.build();
-  }
-
-  @Override
-  protected ImmutableList<TypedPermitRequest> getRequiredQuotas() {
-    return ONE_REST_PRIVATE_POSITION_READ_REQUEST;
-  }
-
-  @Override
-  protected Class<Response> getResponseType() {
-    return Response.class;
-  }
-
-  @NotThreadSafe
-  public static final class Response extends RestResponse<List<_Position>> {}
-}
