+++ bybitlinear-api/src/main/java/io/contek/invoker/bybitlinear/api/rest/user/GetPositionList.java
@@ -19,8 +19,15 @@ import static io.contek.invoker.commons.rest.RestMethod.GET;
 @NotThreadSafe
 public final class GetPositionList extends UserRestRequest<Response> {
 
+  private String symbol;
+
   GetPositionList(IActor actor, RestContext context) {
     super(actor, context);
+  }
+
+  public GetPositionList setSymbol(String symbol) {
+    this.symbol = symbol;
+    return this;
   }
 
   @Override
@@ -35,7 +42,13 @@ public final class GetPositionList extends UserRestRequest<Response> {
 
   @Override
   protected RestParams getParams() {
-    return RestParams.empty();
+    RestParams.Builder builder = RestParams.newBuilder();
+
+    if (symbol != null) {
+      builder.add("symbol", symbol);
+    }
+
+    return builder.build();
   }
 
   @Override
@@ -49,12 +62,5 @@ public final class GetPositionList extends UserRestRequest<Response> {
   }
 
   @NotThreadSafe
-  public static final class Response extends RestResponse<List<Result>> {}
-
-  @NotThreadSafe
-  public static final class Result {
-
-    public boolean is_valid;
-    public _Position data;
-  }
+  public static final class Response extends RestResponse<List<_Position>> {}
 }
